MAINT: Modifies minute bars to use a dict of OHLC ratios
Instead of using a single ratio to scale up pricing for each sid before
writing for bcolz, the writer now accepts a dict mapping each sid to the
ratio to use. This arg is still optional, and we use the existing
OHLC_RATIO (1000) for every sid if no dict is passed.

This allows better handling of futures pricing data, where the required
precision across root symbols is not consistent.